Enhance model and training scripts for improved user message handling and dtype compatibility
- Updated `RewardModel` and `PolicyWithValueHead` classes to ensure the last hidden state is converted to the appropriate dtype for compatibility with the value head.
- Introduced a new utility function `_get_user_message` to streamline the extraction of user messages from input data across multiple training scripts.
- Refactored prompt handling in datasets to utilize the new user message extraction function, improving clarity and consistency in data processing.

+++ 031-reward-modeling/src/model.py
@@ -22,5 +22,6 @@ class RewardModel(nn.Module):
     def forward(self, input_ids: torch.Tensor, attention_mask: torch.Tensor) -> torch.Tensor:
         outputs = self.backbone(input_ids=input_ids, attention_mask=attention_mask)
         # last non-padding token's hidden state
-        last_hidden = outputs.last_hidden_state[:, -1, :]  # (B, H)
-        return self.value_head(last_hidden).squeeze(-1)    # (B,)
+        last_hidden = outputs.last_hidden_state[:, -1, :]           # (B, H)
+        last_hidden = last_hidden.to(self.value_head.weight.dtype)  # backbone may be bf16
+        return self.value_head(last_hidden).squeeze(-1)             # (B,)
